More artifact work.
