+++ packages/packemon-new/src/types.ts
@@ -155,6 +155,7 @@ export interface BuildResult {
 }
 
 export interface Build {
+	declaration: boolean;
 	format: Format;
 	stats?: { size: number };
 }
