--- packages/packemon/src/.eslintrc
@@ -1,9 +1,0 @@
-{
-	"rules": {
-		"no-console": "off",
-		"no-magic-numbers": "off",
-		"no-use-before-define": "off",
-		"sort-keys": "off",
-		"promise/always-return": "off"
-	}
-}
