+++ packages/packemon/src/Package.ts
@@ -1,5 +1,3 @@
-/* eslint-disable require-atomic-updates, no-param-reassign, @typescript-eslint/member-ordering */
-
 import glob from 'fast-glob';
 import fs from 'fs-extra';
 import semver from 'semver';
@@ -38,196 +36,12 @@ import {
 import { TypesArtifact } from './TypesArtifact';
 
 export class Package {
-	readonly artifacts: Artifact[] = [];
-
-	readonly configs: PackageConfig[] = [];
-
-	readonly debug!: Debugger;
-
-	readonly packageJson: PackemonPackage;
-
-	readonly packageJsonPath: Path;
-
-	readonly path: Path;
-
-	readonly project: Project;
-
-	root: boolean = false;
-
-	constructor(project: Project, path: Path, contents: PackemonPackage) {
-		this.project = project;
-		this.path = path;
-		this.packageJsonPath = this.path.append('package.json');
-		this.packageJson = contents;
-		this.debug = createDebugger(['packemon', 'package', this.getSlug()]);
-	}
-
-	addArtifact(artifact: Artifact): Artifact {
-		this.artifacts.push(artifact);
-
-		artifact.startup();
-
-		return artifact;
-	}
-
-	async build(options: BuildOptions, packemonConfig: ConfigFile): Promise<void> {
-		this.debug('Building artifacts');
-
-		// Build artifacts in parallel
-		await Promise.all(
-			this.artifacts.map(async (artifact) => {
-				const start = Date.now();
-
-				try {
-					artifact.state = 'building';
-
-					await artifact.build(options, packemonConfig);
-
-					artifact.state = 'passed';
-				} catch (error: unknown) {
-					artifact.state = 'failed';
-
-					throw error;
-				} finally {
-					artifact.buildResult.time = Date.now() - start;
-				}
-			}),
-		);
-
-		// Add package entry points based on artifacts
-		this.addEntryPoints();
-
-		// Add package `engines` based on artifacts
-		if (options.addEngines) {
-			this.addEngines();
-		}
-
-		// Add package `exports` based on artifacts
-		if (options.addExports) {
-			this.addExports();
-		}
-
-		// Add package `files` whitelist
-		if (options.addFiles) {
-			this.addFiles();
-		}
-
-		// Stamp with a timestamp
-		if (options.stamp) {
-			this.packageJson.release = String(Date.now());
-		}
-
-		// Sync `package.json` in case it was modified
-		await this.syncPackageJson();
-	}
-
-	async cleanup(): Promise<void> {
-		this.debug('Cleaning build artifacts');
-
-		await Promise.all(this.artifacts.map((artifact) => artifact.cleanup()));
-	}
-
-	getName(): string {
-		return this.packageJson.name;
-	}
-
-	getSlug(): string {
-		return this.path.name(true);
-	}
-
-	@Memoize()
-	getSourceFiles(): string[] {
-		const extsWithoutPeriod = EXTENSIONS.map((ext) => ext.slice(1)).join(',');
-
-		return glob
-			.sync(`src/**/*.{${extsWithoutPeriod}}`, {
-				absolute: true,
-				cwd: this.path.path(),
-				onlyFiles: true,
-				// This breaks our own fixtures, so this is hard to test...
-				ignore: process.env.NODE_ENV === 'test' ? [] : EXCLUDE,
-			})
-			.map((file) => new Path(file).path());
-	}
-
 	isComplete(): boolean {
 		return this.artifacts.every((artifact) => artifact.isComplete());
 	}
 
 	isRunning(): boolean {
 		return this.artifacts.some((artifact) => artifact.isRunning());
-	}
-
-	setConfigs(configs: PackemonPackageConfig[]) {
-		configs.forEach((cfg) => {
-			const config = optimal(packemonBlueprint, {
-				name: this.getName(),
-			}).validate(cfg);
-
-			// eslint-disable-next-line complexity
-			toArray(config.platform).forEach((platform) => {
-				let { api, bundle } = config;
-				let formats = config.format ? [config.format] : [];
-
-				switch (platform) {
-					case 'native':
-						formats = formats.filter((format) => (FORMATS_NATIVE as string[]).includes(format));
-
-						if (formats.length === 0) {
-							formats.push(DEFAULT_FORMATS.native);
-						}
-						break;
-
-					case 'node':
-						if (cfg.api === undefined) {
-							api = 'public';
-						}
-
-						if (cfg.bundle === undefined) {
-							bundle = false;
-						}
-
-						formats = formats.filter((format) => (FORMATS_NODE as string[]).includes(format));
-
-						if (formats.length === 0) {
-							formats.push(DEFAULT_FORMATS.node);
-						}
-						break;
-
-					default:
-						formats = formats.filter((format) => (FORMATS_BROWSER as string[]).includes(format));
-
-						if (formats.length === 0) {
-							formats.push(DEFAULT_FORMATS.browser);
-						}
-
-						// Auto-support lib builds for test environments
-						if (formats.includes('esm') && !formats.includes('lib')) {
-							formats.push('lib');
-						}
-
-						if (config.namespace && !formats.includes('umd')) {
-							formats.push('umd');
-						}
-						break;
-				}
-
-				this.configs.push({
-					api,
-					bundle,
-					externals: toArray(config.externals),
-					formats,
-					inputs: config.inputs,
-					namespace: config.namespace,
-					platform,
-					support: config.support,
-				});
-			});
-		});
-	}
-
-	async syncPackageJson() {
-		await fs.writeJson(this.packageJsonPath.path(), this.packageJson, { spaces: 2 });
 	}
 
 	protected addEngines() {
